add: basic language realization

+++ qwerty.sh
@@ -2,21 +2,23 @@
 
 CMAKE="cmake"
 GCC="g++"
-CPP_FLAGS="-Wall -pedantic -Werror -o2"
+CPP_FLAGS="-std=c++14 -Wall -pedantic -Werror -o2"
 
 TARGET_FILE="target.cpp"
-[ "$1" == *.qw ] && RES_FILE="${1/.qw/}" || RES_FILE=$1
+[[ "$1" == *.qw ]] && RES_FILE="${1/.qw/}" || RES_FILE="$1"
 
-rm -rf build
 rm -rf run
 
-mkdir build && cd build
-${CMAKE} -DCMAKE_INSTALL_PREFIX=/media/mark/BSUIR/6_sem/yapis/qwerty_lang/build ../
-make
-cd ..
+if [[ ! -f 'build/qwerty_lang' ]]; then
+    rm -rf build
+    mkdir build && cd build
+    ${CMAKE} -DCMAKE_INSTALL_PREFIX=/media/mark/BSUIR/6_sem/yapis/qwerty_lang/build ../
+    make
+    cd ..
+fi
+
 ./build/qwerty_lang $1 ${TARGET_FILE}
 
 mkdir run && cd run
-cd run
 mv ../${TARGET_FILE} ./
 ${GCC} ${CPP_FLAGS} ${TARGET_FILE} -o ${RES_FILE}
